# internal/httpserver/openapi.yaml
openapi: 3.0.3
info:
  title: GTFS Mobile API
  version: 1.0.0
  description: |
    Thin Go layer over Supabase GTFS static + realtime data.
    All responses follow the envelope `{ status, data, meta? }`.
    `status.mode` becomes `static_only` if the GTFS realtime feed has not synced recently.
servers:
  - url: http://localhost:8080
  - url: https://localhost:8443
...
paths:
...
  /v1/auth/signup:
    post:
      tags: [Auth]
      summary: Sign up a user with Supabase Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRequest"
      responses:
        "200":
          description: Signup response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/auth/login:
    post:
      tags: [Auth]
      summary: Login a user with Supabase Auth password grant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRequest"
      responses:
        "200":
          description: Login response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid credentials or request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/auth/verify-otp:
    post:
      tags: [Auth]
      summary: Verify email OTP for signup or email flows
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyOTPRequest"
      responses:
        "200":
          description: Verification response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid token/type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/gtfs/stops:
    get:
      tags: [Stops]
      summary: Search stops by text and/or geo
      parameters:
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: lat
          schema: { type: number, format: double }
        - in: query
          name: lon
          schema: { type: number, format: double }
        - in: query
          name: radius
          description: Radius in meters
          schema: { type: integer, minimum: 1 }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 200 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StopSummaryListResponse" }
  /v1/gtfs/stops/{stopId}:
    get:
      tags: [Stops]
      summary: Get stop details
      parameters:
        - in: path
          name: stopId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StopResponse" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /v1/gtfs/stops/{stopId}/departures:
    get:
      tags: [Stops]
      summary: Departures at a stop
      parameters:
        - in: path
          name: stopId
          required: true
          schema: { type: string }
        - in: query
          name: at
          description: ISO 8601 timestamp (defaults to now)
          schema: { type: string, format: date-time }
        - in: query
          name: windowMinutes
          schema: { type: integer, minimum: 1, maximum: 360 }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 50 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DeparturesResponse" }
  /v1/gtfs/routes:
    get:
      tags: [Routes]
      summary: Search routes
      parameters:
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: type
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 500 }
        - in: query
          name: page
          schema: { type: integer, minimum: 1 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RoutesResponse" }
  /v1/gtfs/routes/{routeId}:
    get:
      tags: [Routes]
      summary: Get route details
      parameters:
        - in: path
          name: routeId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RouteResponse" }
        "404":
          description: Not found
  /v1/gtfs/routes/{routeId}/stops:
    get:
      tags: [Routes]
      summary: Stops served by a route
      parameters:
        - in: path
          name: routeId
          required: true
          schema: { type: string }
        - in: query
          name: directionId
          schema: { type: string, enum: ["0", "1"] }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StopSummaryListResponse" }
  /v1/gtfs/routes/{routeId}/directions:
    get:
      tags: [Routes]
      summary: Directions available for a route
      parameters:
        - in: path
          name: routeId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RouteDirectionsResponse" }
  /v1/gtfs/routes/{routeId}/shape:
    get:
      tags: [Routes]
      summary: Polyline points for most common route shape
      parameters:
        - in: path
          name: routeId
          required: true
          schema: { type: string }
        - in: query
          name: directionId
          schema: { type: string, enum: ["0", "1"] }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RouteShapeResponse" }
        "404":
          description: Not found
  /v1/gtfs/trips:
    get:
      tags: [Trips]
      summary: List trips by route and service date
      parameters:
        - in: query
          name: routeId
          schema: { type: string }
        - in: query
          name: serviceDate
          description: YYYY-MM-DD (defaults to today)
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 1000 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TripsResponse" }
  /v1/gtfs/trips/{tripId}:
    get:
      tags: [Trips]
      summary: Get trip details
      parameters:
        - in: path
          name: tripId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TripResponse" }
        "404":
          description: Not found
  /v1/gtfs/trips/{tripId}/stop-times:
    get:
      tags: [Trips]
      summary: Stop times for a trip
      parameters:
        - in: path
          name: tripId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TripStopTimesResponse" }
  /v1/gtfs/calendar/service:
    get:
      tags: [Calendar]
      summary: Active services on a date
      parameters:
        - in: query
          name: date
          description: YYYY-MM-DD (defaults to today)
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ServiceCalendarResponse" }
  /v1/gtfs/calendar/exceptions:
    get:
      tags: [Calendar]
      summary: Calendar exceptions in a date range
      parameters:
        - in: query
          name: from
          required: true
          schema: { type: string }
        - in: query
          name: to
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CalendarExceptionsResponse" }
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
    AuthRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
        full_name: { type: string, description: "Required for signup endpoint" }
        metadata:
          type: object
          additionalProperties: true
    VerifyOTPRequest:
      type: object
      required: [token, type]
      properties:
        email: { type: string, format: email }
        token: { type: string, description: "OTP or token from email" }
        type:
          type: string
          enum: [signup, email, recovery, magiclink, email_change]
    AuthResponse:
      type: object
      properties:
        access_token: { type: string }
        refresh_token: { type: string }
        token_type: { type: string }
        expires_in: { type: integer }
        user:
          type: object
          additionalProperties: true
    StopSummaryListResponse:
      type: object
      properties:
        status: { $ref: "#/components/schemas/APIStatus" }
        data:
          type: array
          items: { $ref: "#/components/schemas/StopSummary" }
        meta: { $ref: "#/components/schemas/Meta" }
    StopResponse:
      type: object
      properties:
        status: { $ref: "#/components/schemas/APIStatus" }
        data: { $ref: "#/components/schemas/Stop" }
    DeparturesResponse:
      type: object
      properties:
        status: { $ref: "#/components/schemas/APIStatus" }
        data:
          type: array
          items: { $ref: "#/components/schemas/Departure" }
        meta: { $ref: "#/components/schemas/Meta" }
    RoutesResponse:
      type: object
      properties:
        status: { $ref: "#/components/schemas/APIStatus" }
        data:
          type: array
          items: { $ref: "#/components/schemas/Route" }
        meta: { $ref: "#/components/schemas/Meta" }
    RouteResponse:
      type: object
      properties:
        status: { $ref: "#/components/schemas/APIStatus" }
        data: { $ref: "#/components/schemas/Route" }
    RouteShapeResponse:
      type: object
      properties:
        status: { $ref: "#/components/schemas/APIStatus" }
        data: { $ref: "#/components/schemas/RouteShape" }
        meta: { $ref: "#/components/schemas/Meta" }
    RouteDirectionsResponse:
      type: object
      properties:
        status: { $ref: "#/components/schemas/APIStatus" }
        data:
          type: array
          items: { $ref: "#/components/schemas/RouteDirection" }
        meta: { $ref: "#/components/schemas/Meta" }
    TripsResponse:
      type: object
      properties:
        status: { $ref: "#/components/schemas/APIStatus" }
        data:
          type: array
          items: { $ref: "#/components/schemas/Trip" }
        meta: { $ref: "#/components/schemas/Meta" }
    TripResponse:
      type: object
      properties:
        status: { $ref: "#/components/schemas/APIStatus" }
        data: { $ref: "#/components/schemas/Trip" }
...